move field adapters into folder
