+++ test/field_transfer_example.cpp
@@ -1,7 +1,7 @@
 #include <Omega_h_build.hpp>
 #include <Omega_h_vtk.hpp>
 #include <pcms/transfer_field.h>
-#include <pcms/omega_h_field.h>
+#include "pcms/adapter/omega_h/omega_h_field.h"
 // for transfer operation dummy test
 #include <pcms.h>
 #include <chrono>
