+++ test/test_proxy_coupling.cpp
@@ -6,7 +6,7 @@
 #include <Omega_h_for.hpp>
 #include <redev_variant_tools.h>
 #include "test_support.h"
-#include <pcms/omega_h_field.h>
+#include "pcms/adapter/omega_h/omega_h_field.h"
 #include <chrono>
 #include <thread>
