+++ test/test_proxy_coupling_xgc_server.cpp
@@ -4,8 +4,8 @@
 #include <Omega_h_file.hpp>
 #include <Omega_h_for.hpp>
 #include "test_support.h"
-#include <pcms/omega_h_field.h>
-#include <pcms/xgc_field_adapter.h>
+#include "pcms/adapter/omega_h/omega_h_field.h"
+#include "pcms/adapter/xgc/xgc_field_adapter.h"
 
 using pcms::ConstructRCFromOmegaHMesh;
 using pcms::Copy;
