+++ test/xgc_n0_coupling_server.cpp
@@ -3,8 +3,8 @@
 #include <Omega_h_file.hpp>
 #include <Omega_h_for.hpp>
 #include "test_support.h"
-#include <pcms/omega_h_field.h>
-#include <pcms/xgc_field_adapter.h>
+#include "pcms/adapter/omega_h/omega_h_field.h"
+#include "pcms/adapter/xgc/xgc_field_adapter.h"
 #include <chrono>
 
 using pcms::Copy;
