+++ tools/XgcRCfromOsh.cpp
@@ -1,5 +1,5 @@
 #include <Omega_h_file.hpp>
-#include <pcms/xgc_reverse_classification.h>
+#include "pcms/adapter/xgc/xgc_reverse_classification.h"
 #include "pcms/print.h"
 #include <Omega_h_mesh.hpp>
 #include <Omega_h_tag.hpp>
